requirements.txt
certifi==2019.11.28
cycler==0.10.0
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.5
olefile==0.46
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
tornado==6.0.3
wincertstore==0.2
